patch: Update source to supported DiscordUser integrated economy manager
