+++ src/services/EconomyHandler.ts
@@ -1,8 +1,8 @@
 import { ChannelType } from 'discord.js';
 import { client } from '../index';
 import { prisma } from '../utils/DatabaseManager';
-import { grantPoints } from '../DBUtils/EconomyManager';
 import { noPointsChannel } from '../config';
+import { DiscordUser } from '../ManagerUtils/DiscordUser';
 
 
 // Grant currency based on chats
@@ -16,5 +16,5 @@ client.on('messageCreate', async (message) => {
     // Prevent points from being awarded to blacklisted channels
     if(noPointsChannel.find((c)=> c === message.channel.id)) return;
     // Grant the user the points
-    await grantPoints(message.author.id);
+    await (new DiscordUser(message.author)).economy.grantPoints();
 })
